+++ changelog.md
@@ -18,6 +18,16 @@ Integration of mobile payment apps is impossible due to privacy reasons maybe pu
 
 ## Changelog
 
+### 3.1.2 bundle 29 (27.08.2025)
+
+- fix categories widget
+
+### 3.1.1 bundle 28 (26.08.2025)
+
+- fix list widget
+- add widget names
+- allow picking pdf files in scanner
+
 ### 3.1.0 bundle 27 (24.08.2025)
 
 - scan function to extract info from receiptand automatically assign sum to bucket